+++ nqp/ui_elements/tailored/unit_stats_window.py
@@ -3,10 +3,8 @@ from __future__ import annotations
 from typing import TYPE_CHECKING
 
 import pygame
-from pygame import SRCALPHA
 
 from nqp.command.unit import Unit
-from nqp.core.components import Stats
 from nqp.core.constants import (
     DEFAULT_IMAGE_SIZE,
     FontType,
@@ -17,10 +15,9 @@ from nqp.core.constants import (
 )
 from nqp.ui_elements.generic.ui_frame import UIFrame
 from nqp.ui_elements.generic.ui_window import UIWindow
+from nqp.world_elements.entity_components import Stats
 
 if TYPE_CHECKING:
-    from typing import Tuple
-
     from nqp.core.game import Game
